fix: add existence checks before downloading jar files in scripts

+++ .bin/jar/floodgate.sh
@@ -1,2 +1,4 @@
-curl -L -o Floodgate-Spigot.jar \
-  "https://download.geysermc.org/v2/projects/floodgate/versions/latest/builds/latest/downloads/spigot"
+if [ ! -f "./plugins/Floodgate-Spigot.jar" ] && [ ! -f "./Floodgate-Spigot.jar" ]; then
+  curl -L -o Floodgate-Spigot.jar \
+    "https://download.geysermc.org/v2/projects/floodgate/versions/latest/builds/latest/downloads/spigot"
+fi
